requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0

# LangChain essentials
langchain==0.1.0
langchain-core==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
langchain-ollama==0.0.1

# LLM APIs
openai==1.6.0
anthropic==0.8.0

# Frameworks
crewai==0.1.0
pyautogen==0.2.0

# Utilities
jinja2==3.1.2
tiktoken==0.5.2